SWT token cleanup
